# GoWebApi/api/gen/pbauth/refresh.pb.go
// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.35.2
// 	protoc        v5.28.3
// source: proto/pbauth/refresh.proto

package pbauth

import (
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	reflect "reflect"
	sync "sync"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

// 请求-刷新Jwt
type RefreshRequest struct {
	state         protoimpl.MessageState
	sizeCache     protoimpl.SizeCache
	unknownFields protoimpl.UnknownFields
}

func (x *RefreshRequest) Reset() {
	*x = RefreshRequest{}
	mi := &file_proto_pbauth_refresh_proto_msgTypes[0]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *RefreshRequest) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*RefreshRequest) ProtoMessage() {}

func (x *RefreshRequest) ProtoReflect() protoreflect.Message {
	mi := &file_proto_pbauth_refresh_proto_msgTypes[0]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use RefreshRequest.ProtoReflect.Descriptor instead.
func (*RefreshRequest) Descriptor() ([]byte, []int) {
	return file_proto_pbauth_refresh_proto_rawDescGZIP(), []int{0}
}

// 响应-刷新Jwt
type RefreshResponse struct {
	state         protoimpl.MessageState
	sizeCache     protoimpl.SizeCache
	unknownFields protoimpl.UnknownFields

	Token string `protobuf:"bytes,1,opt,name=token,proto3" json:"token,omitempty"`
}

func (x *RefreshResponse) Reset() {
	*x = RefreshResponse{}
	mi := &file_proto_pbauth_refresh_proto_msgTypes[1]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *RefreshResponse) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*RefreshResponse) ProtoMessage() {}

func (x *RefreshResponse) ProtoReflect() protoreflect.Message {
	mi := &file_proto_pbauth_refresh_proto_msgTypes[1]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use RefreshResponse.ProtoReflect.Descriptor instead.
func (*RefreshResponse) Descriptor() ([]byte, []int) {
	return file_proto_pbauth_refresh_proto_rawDescGZIP(), []int{1}
}

func (x *RefreshResponse) GetToken() string {
	if x != nil {
		return x.Token
	}
	return ""
}

var File_proto_pbauth_refresh_proto protoreflect.FileDescriptor

var file_proto_pbauth_refresh_proto_rawDesc = []byte{
	0x0a, 0x1a, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x70, 0x62, 0x61, 0x75, 0x74, 0x68, 0x2f, 0x72,
	0x65, 0x66, 0x72, 0x65, 0x73, 0x68, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x06, 0x70, 0x62,
	0x61, 0x75, 0x74, 0x68, 0x22, 0x10, 0x0a, 0x0e, 0x52, 0x65, 0x66, 0x72, 0x65, 0x73, 0x68, 0x52,
	0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x22, 0x27, 0x0a, 0x0f, 0x52, 0x65, 0x66, 0x72, 0x65, 0x73,
	0x68, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x14, 0x0a, 0x05, 0x74, 0x6f, 0x6b,
	0x65, 0x6e, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x74, 0x6f, 0x6b, 0x65, 0x6e, 0x42,
	0x1e, 0x5a, 0x1c, 0x77, 0x65, 0x62, 0x61, 0x70, 0x69, 0x2f, 0x61, 0x70, 0x69, 0x2f, 0x67, 0x65,
	0x6e, 0x2f, 0x70, 0x62, 0x61, 0x75, 0x74, 0x68, 0x3b, 0x70, 0x62, 0x61, 0x75, 0x74, 0x68, 0x62,
	0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
}

var (
	file_proto_pbauth_refresh_proto_rawDescOnce sync.Once
	file_proto_pbauth_refresh_proto_rawDescData = file_proto_pbauth_refresh_proto_rawDesc
)

func file_proto_pbauth_refresh_proto_rawDescGZIP() []byte {
	file_proto_pbauth_refresh_proto_rawDescOnce.Do(func() {
		file_proto_pbauth_refresh_proto_rawDescData = protoimpl.X.CompressGZIP(file_proto_pbauth_refresh_proto_rawDescData)
	})
	return file_proto_pbauth_refresh_proto_rawDescData
}

var file_proto_pbauth_refresh_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
var file_proto_pbauth_refresh_proto_goTypes = []any{
	(*RefreshRequest)(nil),  // 0: pbauth.RefreshRequest
	(*RefreshResponse)(nil), // 1: pbauth.RefreshResponse
}
var file_proto_pbauth_refresh_proto_depIdxs = []int32{
	0, // [0:0] is the sub-list for method output_type
	0, // [0:0] is the sub-list for method input_type
	0, // [0:0] is the sub-list for extension type_name
	0, // [0:0] is the sub-list for extension extendee
	0, // [0:0] is the sub-list for field type_name
}

func init() { file_proto_pbauth_refresh_proto_init() }
func file_proto_pbauth_refresh_proto_init() {
	if File_proto_pbauth_refresh_proto != nil {
		return
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: file_proto_pbauth_refresh_proto_rawDesc,
			NumEnums:      0,
			NumMessages:   2,
			NumExtensions: 0,
			NumServices:   0,
		},
		GoTypes:           file_proto_pbauth_refresh_proto_goTypes,
		DependencyIndexes: file_proto_pbauth_refresh_proto_depIdxs,
		MessageInfos:      file_proto_pbauth_refresh_proto_msgTypes,
	}.Build()
	File_proto_pbauth_refresh_proto = out.File
	file_proto_pbauth_refresh_proto_rawDesc = nil
	file_proto_pbauth_refresh_proto_goTypes = nil
	file_proto_pbauth_refresh_proto_depIdxs = nil
}