# GoWebApi/api/gen/pbauth/register.pb.go
// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.35.2
// 	protoc        v5.28.3
// source: proto/pbauth/register.proto

package pbauth

import (
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	reflect "reflect"
	sync "sync"
	pbuser "webapi/api/gen/pbuser"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

// 请求-注册
type RegisterRequest struct {
	state         protoimpl.MessageState
	sizeCache     protoimpl.SizeCache
	unknownFields protoimpl.UnknownFields

	// 用户邮箱
	Email string `protobuf:"bytes,1,opt,name=email,proto3" json:"email,omitempty" binding:"required,email"` // @gotags: binding:"required,email"
	// 用户密码
	Password string `protobuf:"bytes,2,opt,name=password,proto3" json:"password,omitempty" binding:"required,min=6,max=20"` // @gotags: binding:"required,min=6,max=20"
	// 用户名
	Uname string `protobuf:"bytes,3,opt,name=uname,proto3" json:"uname,omitempty" binding:"required,min=2,max=30"` // @gotags: binding:"required,min=2,max=30"
	// 用户年龄
	Age uint32 `protobuf:"varint,4,opt,name=age,proto3" json:"age,omitempty" binding:"required,gte=0,lte=200"` // @gotags: binding:"required,gte=0,lte=200"
}

func (x *RegisterRequest) Reset() {
	*x = RegisterRequest{}
	mi := &file_proto_pbauth_register_proto_msgTypes[0]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *RegisterRequest) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*RegisterRequest) ProtoMessage() {}

func (x *RegisterRequest) ProtoReflect() protoreflect.Message {
	mi := &file_proto_pbauth_register_proto_msgTypes[0]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use RegisterRequest.ProtoReflect.Descriptor instead.
func (*RegisterRequest) Descriptor() ([]byte, []int) {
	return file_proto_pbauth_register_proto_rawDescGZIP(), []int{0}
}

func (x *RegisterRequest) GetEmail() string {
	if x != nil {
		return x.Email
	}
	return ""
}

func (x *RegisterRequest) GetPassword() string {
	if x != nil {
		return x.Password
	}
	return ""
}

func (x *RegisterRequest) GetUname() string {
	if x != nil {
		return x.Uname
	}
	return ""
}

func (x *RegisterRequest) GetAge() uint32 {
	if x != nil {
		return x.Age
	}
	return 0
}

// 响应-注册
type RegisterResponse struct {
	state         protoimpl.MessageState
	sizeCache     protoimpl.SizeCache
	unknownFields protoimpl.UnknownFields

	Token string       `protobuf:"bytes,1,opt,name=token,proto3" json:"token,omitempty"`
	User  *pbuser.User `protobuf:"bytes,2,opt,name=user,proto3" json:"user,omitempty"`
}

func (x *RegisterResponse) Reset() {
	*x = RegisterResponse{}
	mi := &file_proto_pbauth_register_proto_msgTypes[1]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *RegisterResponse) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*RegisterResponse) ProtoMessage() {}

func (x *RegisterResponse) ProtoReflect() protoreflect.Message {
	mi := &file_proto_pbauth_register_proto_msgTypes[1]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use RegisterResponse.ProtoReflect.Descriptor instead.
func (*RegisterResponse) Descriptor() ([]byte, []int) {
	return file_proto_pbauth_register_proto_rawDescGZIP(), []int{1}
}

func (x *RegisterResponse) GetToken() string {
	if x != nil {
		return x.Token
	}
	return ""
}

func (x *RegisterResponse) GetUser() *pbuser.User {
	if x != nil {
		return x.User
	}
	return nil
}

var File_proto_pbauth_register_proto protoreflect.FileDescriptor

var file_proto_pbauth_register_proto_rawDesc = []byte{
	0x0a, 0x1b, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x70, 0x62, 0x61, 0x75, 0x74, 0x68, 0x2f, 0x72,
	0x65, 0x67, 0x69, 0x73, 0x74, 0x65, 0x72, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x06, 0x70,
	0x62, 0x61, 0x75, 0x74, 0x68, 0x1a, 0x17, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x70, 0x62, 0x75,
	0x73, 0x65, 0x72, 0x2f, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0x6b,
	0x0a, 0x0f, 0x52, 0x65, 0x67, 0x69, 0x73, 0x74, 0x65, 0x72, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73,
	0x74, 0x12, 0x14, 0x0a, 0x05, 0x65, 0x6d, 0x61, 0x69, 0x6c, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09,
	0x52, 0x05, 0x65, 0x6d, 0x61, 0x69, 0x6c, 0x12, 0x1a, 0x0a, 0x08, 0x70, 0x61, 0x73, 0x73, 0x77,
	0x6f, 0x72, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x70, 0x61, 0x73, 0x73, 0x77,
	0x6f, 0x72, 0x64, 0x12, 0x14, 0x0a, 0x05, 0x75, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x03, 0x20, 0x01,
	0x28, 0x09, 0x52, 0x05, 0x75, 0x6e, 0x61, 0x6d, 0x65, 0x12, 0x10, 0x0a, 0x03, 0x61, 0x67, 0x65,
	0x18, 0x04, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x03, 0x61, 0x67, 0x65, 0x22, 0x4a, 0x0a, 0x10, 0x52,
	0x65, 0x67, 0x69, 0x73, 0x74, 0x65, 0x72, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12,
	0x14, 0x0a, 0x05, 0x74, 0x6f, 0x6b, 0x65, 0x6e, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05,
	0x74, 0x6f, 0x6b, 0x65, 0x6e, 0x12, 0x20, 0x0a, 0x04, 0x75, 0x73, 0x65, 0x72, 0x18, 0x02, 0x20,
	0x01, 0x28, 0x0b, 0x32, 0x0c, 0x2e, 0x70, 0x62, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x55, 0x73, 0x65,
	0x72, 0x52, 0x04, 0x75, 0x73, 0x65, 0x72, 0x42, 0x1e, 0x5a, 0x1c, 0x77, 0x65, 0x62, 0x61, 0x70,
	0x69, 0x2f, 0x61, 0x70, 0x69, 0x2f, 0x67, 0x65, 0x6e, 0x2f, 0x70, 0x62, 0x61, 0x75, 0x74, 0x68,
	0x3b, 0x70, 0x62, 0x61, 0x75, 0x74, 0x68, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
}

var (
	file_proto_pbauth_register_proto_rawDescOnce sync.Once
	file_proto_pbauth_register_proto_rawDescData = file_proto_pbauth_register_proto_rawDesc
)

func file_proto_pbauth_register_proto_rawDescGZIP() []byte {
	file_proto_pbauth_register_proto_rawDescOnce.Do(func() {
		file_proto_pbauth_register_proto_rawDescData = protoimpl.X.CompressGZIP(file_proto_pbauth_register_proto_rawDescData)
	})
	return file_proto_pbauth_register_proto_rawDescData
}

var file_proto_pbauth_register_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
var file_proto_pbauth_register_proto_goTypes = []any{
	(*RegisterRequest)(nil),  // 0: pbauth.RegisterRequest
	(*RegisterResponse)(nil), // 1: pbauth.RegisterResponse
	(*pbuser.User)(nil),      // 2: pbuser.User
}
var file_proto_pbauth_register_proto_depIdxs = []int32{
	2, // 0: pbauth.RegisterResponse.user:type_name -> pbuser.User
	1, // [1:1] is the sub-list for method output_type
	1, // [1:1] is the sub-list for method input_type
	1, // [1:1] is the sub-list for extension type_name
	1, // [1:1] is the sub-list for extension extendee
	0, // [0:1] is the sub-list for field type_name
}

func init() { file_proto_pbauth_register_proto_init() }
func file_proto_pbauth_register_proto_init() {
	if File_proto_pbauth_register_proto != nil {
		return
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: file_proto_pbauth_register_proto_rawDesc,
			NumEnums:      0,
			NumMessages:   2,
			NumExtensions: 0,
			NumServices:   0,
		},
		GoTypes:           file_proto_pbauth_register_proto_goTypes,
		DependencyIndexes: file_proto_pbauth_register_proto_depIdxs,
		MessageInfos:      file_proto_pbauth_register_proto_msgTypes,
	}.Build()
	File_proto_pbauth_register_proto = out.File
	file_proto_pbauth_register_proto_rawDesc = nil
	file_proto_pbauth_register_proto_goTypes = nil
	file_proto_pbauth_register_proto_depIdxs = nil
}