# GoWebApi/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "API接口",
        "contact": {}
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "登录用户",
                "parameters": [
                    {
                        "description": "请求body参数",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/pbauth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pbauth.LoginResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "刷新Jwt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token令牌",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pbauth.RefreshResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "注册用户",
                "parameters": [
                    {
                        "description": "请求body参数",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/pbauth.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pbauth.RegisterResponse"
                        }
                    }
                }
            }
        },
        "/api/user/info": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token令牌",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pbuser.InfoRequest"
                        }
                    }
                }
            }
        },
        "/api/user/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "每页页数(数量)",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "当前页数",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pbuser.ListResponse"
                        }
                    }
                }
            }
        },
        "/api/user/update": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "修改用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "请求body参数",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/pbuser.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pbuser.UpdateResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "pbauth.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6
                }
            }
        },
        "pbauth.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/pbuser.User"
                }
            }
        },
        "pbauth.RefreshResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "pbauth.RegisterRequest": {
            "type": "object",
            "required": [
                "age",
                "email",
                "password",
                "uname"
            ],
            "properties": {
                "age": {
                    "description": "用户年龄",
                    "type": "integer",
                    "maximum": 200,
                    "minimum": 0
                },
                "email": {
                    "description": "用户邮箱",
                    "type": "string"
                },
                "password": {
                    "description": "用户密码",
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6
                },
                "uname": {
                    "description": "用户名",
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 2
                }
            }
        },
        "pbauth.RegisterResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/pbuser.User"
                }
            }
        },
        "pbuser.InfoRequest": {
            "type": "object"
        },
        "pbuser.ListResponse": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pbuser.User"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "pbuser.UpdateRequest": {
            "type": "object",
            "required": [
                "login_at",
                "uname"
            ],
            "properties": {
                "login_at": {
                    "description": "登录时间",
                    "type": "string"
                },
                "uname": {
                    "description": "用户名",
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 2
                }
            }
        },
        "pbuser.UpdateResponse": {
            "type": "object"
        },
        "pbuser.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "email": {
                    "description": "用户邮箱",
                    "type": "string"
                },
                "uid": {
                    "description": "用户ID",
                    "type": "integer"
                },
                "uname": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        }
    }
}